bot/pyproject.toml
[project]
name = "bdo_daily_bot"
version = "4.0.0"
description = "Discord Bot for organizing and collecting daily sea raids for the game Black Desert Online[RU](BDO)"
readme = "README.md"
requires-python = ">=3.11"
keywords = ["bdo", "discord", "bot"]
authors = [{name = "Andrei Zaneuski", email = "zanevskiyandrey@gmail.com"}]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Programming Language :: Python :: 3.11",
  "Topic :: Software Development :: Version Control :: Git",
]

dependencies = [
  "discord-py-interactions",
  "pyyaml",
  "motor",
  "numpy",
  "opencv-python",
  "Pillow",
  "pymongo",
]


[project.optional-dependencies]
dev = [
    "black",
    "mypy",
    "pip-tools",
    "pre-commit",
    "pylint"
]

[tool.setuptools.package-data]
"bdo_daily_bot.settings" = ["font_table.ttf"]

[tool.setuptools.packages.find]
include = ["bdo_daily_bot*"]

[project.scripts]
run_bot = "bdo_daily_bot.bot:run_bot"

[project.urls]
repository = "https://github.com/Gliger13/bdo_daily_bot"
